refactor: common/ui Header, icon 공통화
