+++ src/page/cart/index.tsx
@@ -6,6 +6,7 @@ import LeftSection from "./left-section";
 import RightSection from "./right-section";
 
 import { useCart } from "hooks/cart";
+import { Header } from "common/header";
 
 const CartContent = () => {
   const { data: carts, isLoading, isError } = useCart();
@@ -24,11 +25,7 @@ const CartContent = () => {
 
   return (
     <section className="cart-section">
-      <header className="flex-col-center mt-20">
-        <h2 className="cart-section__title">장바구니</h2>
-        <hr className="divide-line mt-20" />
-      </header>
-
+      <Header title={'장바구니'}/>
       <div className="flex">
         <LeftSection carts={carts}/>
         <RightSection />
